Move CPUShares adjustment out of daemon.adaptContainerSettings()
Commit ed39fbeb2ad3959f37cf6c16aaf30aacb3292817 added a version-check to the
CPUShares property, which was introduced in API 1.19.

In an attempt to cleanup code, ed39fbeb2ad3959f37cf6c16aaf30aacb3292817 kept
the "version" check in the API, but added a boolean to enable/disable this
feature all the way to the daemon.

This patch removes the boolean, and puts the adjustment back into a separate
method, to better clarify it's for backward compatibility.

Also removes the adjustment for Solaris, because Solaris does not support
API 1.19, so does not require these.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>