Fix shutdown-timeout configuration reload
When live-reloading the daemon configuration after *removing* "shutdown-timeout"
from the daemon.json configuration file, the previously configured value was
not reset to its default.

This change removes the check if the configuration was set, and if not, reset
to the default value.

Before this change

    echo '{ "shutdown-timeout" : 300 }' > /etc/docker/daemon.json
    pkill -HUP dockerd

Check the daemon logs (truncated for readability):

    INFO[2018-03-20T12:30:17.062096063Z] Reloaded configuration: {"shutdown-timeout":300}

Remove the timeout from the configuration file, and reload the daemon:

    echo '{}' > /etc/docker/daemon.json
    pkill -HUP dockerd

Check the daemon logs (truncated for readability), and notice that the
value is not reset:

    INFO[2018-03-20T12:30:17.062096063Z] Reloaded configuration: {"shutdown-timeout":300}

With this patch applied

    echo '{}' > /etc/docker/daemon.json
    pkill -HUP dockerd

Check the daemon logs (truncated for readability), and notice that the
shutdown-timeout is reset to its default value:

    INFO[2018-03-20T12:45:08.353388357Z] Reloaded configuration: {"shutdown-timeout":15}

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>