Fix "labels" configuration reload
When live-reloading the daemon configuration after *removing* "labels" from
the daemon.json configuration file, the previously configured value was not
reset to its default.

This change removes the check if the configuration was set, and if not, reset
to the default value.

Before this change:

    echo '{ "labels" : [ "foo=bar", "bar=baz" ] }' > /etc/docker/daemon.json
    pkill -HUP dockerd

    docker info --format '{{json .Labels }}'
    ["foo=bar","bar=baz"]

Removing the labels from the configuration, and reloading has no effect:

    echo '{}' > /etc/docker/daemon.json
    pkill -HUP dockerd

    docker info --format '{{json .Labels }}'
    ["foo=bar","bar=baz"]

With this patch applied:

    echo '{ "labels" : [ "foo=bar", "bar=baz" ] }' > /etc/docker/daemon.json
    pkill -HUP dockerd

    docker info --format '{{json .Labels }}'
    ["foo=bar","bar=baz"]

Updating the configuration removes the labels:

    echo '{}' > /etc/docker/daemon.json
    pkill -HUP dockerd

    docker info --format '{{json .Labels }}'
    []

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>