errdefs: remove dependency on grpc, containerd, and distribution
The errdefs package is used both to generate errors, but also has utilities
to convert errors originating from subsystems into a suitable HTTP error.

The API should always return a suitable HTTP error, and Conversion of errors
originating from subsystems will (generally) only be needed in the API server
(not the client).

This patch prevents the client from depending on grpc, containerd and distribution,
by moving the code to convert those errors to a separate "adapters" package.

The api/server/httputils.GetHTTPErrorStatusCode was un-deprecated, and modified
to include the default set of adapters.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>