WIP: don't force-kill container when cancelling context
Commit 90de570cfa5f6de0b02bdf2e794627fbf5b0dfc8 passed through the request
context to daemon.ContainerStop(). As a result, cancelling the context would
cancel the "graceful" stop of the container, and would proceed with forcefully
killing the container.

This patch fixes the context-cancellation to cancel the stop.

TODO: add a `?wait=false` option, otherwise this this is a change in behavior (as
older versions expected the stop to continue running in the background, even if the
client detached).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>