api: Deprecate ContainerConfig.MacAddress
Having a sandbox/container-wide MacAddress field makes little since a
container can be connected to multiple networks at the same time. This
field is an artefact of old times where a container could be connected
to a single network only. As the previous commit added a way to specify
per-endpoint mac address, this field is now useless and thus deprecated.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>